@@ -18,15 +18,13 @@
     "not op_mini all"
   ],
   "dependencies": {
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.2.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^7.2.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "imagemin": "^7.0.0",
-    "imagemin-jpegtran": "^6.0.0",
-    "imagemin-optipng": "^7.1.0",
-    "imagemin-svgo": "^7.0.0"
+    "sharp": "^0.33.2",
+    "svgo": "^3.2.0"
   }
 }
